fix: [N19]: Redundant code

+++ packages/core/contracts/oracle/implementation/VotingV2.sol
@@ -1008,8 +1008,8 @@ contract VotingV2 is
         if (resolutionPrice == 1e18) {
             // Delete the price requests associated with the spam.
             for (uint256 i = 0; i < spamDeletionProposals[proposalId].spamRequestIndices.length; i = unsafe_inc(i)) {
-                uint64 startIndex = uint64(spamDeletionProposals[proposalId].spamRequestIndices[uint256(i)][0]);
-                uint64 endIndex = uint64(spamDeletionProposals[proposalId].spamRequestIndices[uint256(i)][1]);
+                uint64 startIndex = spamDeletionProposals[proposalId].spamRequestIndices[uint256(i)][0];
+                uint64 endIndex = spamDeletionProposals[proposalId].spamRequestIndices[uint256(i)][1];
                 for (uint256 j = startIndex; j <= endIndex; j++) {
                     bytes32 requestId = priceRequestIds[j];
                     // Remove from pendingPriceRequests.
